Shift a bunch more string-related functions to strutils
