+++ mitmproxy/contentviews.py
@@ -37,7 +37,6 @@ from netlib import http
 from netlib import odict
 from netlib.http import url
 from netlib import strutils
-import netlib.utils
 
 try:
     import pyamf
@@ -130,11 +129,11 @@ class ViewAuto(View):
             ct = "%s/%s" % (ct[0], ct[1])
             if ct in content_types_map:
                 return content_types_map[ct][0](data, **metadata)
-            elif mitmproxy.utils.isXML(data):
+            elif strutils.isXML(data):
                 return get("XML")(data, **metadata)
         if metadata.get("query"):
             return get("Query")(data, **metadata)
-        if data and mitmproxy.utils.isMostlyBin(data):
+        if data and strutils.isMostlyBin(data):
             return get("Hex")(data)
         if not data:
             return "No content", []
@@ -157,7 +156,7 @@ class ViewHex(View):
 
     @staticmethod
     def _format(data):
-        for offset, hexa, s in netlib.utils.hexdump(data):
+        for offset, hexa, s in strutils.hexdump(data):
             yield [
                 ("offset", offset + " "),
                 ("text", hexa + "   "),
@@ -227,7 +226,7 @@ class ViewHTML(View):
     content_types = ["text/html"]
 
     def __call__(self, data, **metadata):
-        if mitmproxy.utils.isXML(data):
+        if strutils.isXML(data):
             parser = lxml.etree.HTMLParser(
                 strip_cdata=True,
                 remove_blank_text=True
