+++ pathod/log.py
@@ -2,9 +2,6 @@ import datetime
 
 import six
 
-import netlib.utils
-import netlib.tcp
-import netlib.http
 from netlib import strutils
 
 TIMEFMT = '%d-%m-%y %H:%M:%S'
@@ -63,7 +60,7 @@ class LogCtx(object):
 
     def dump(self, data, hexdump):
         if hexdump:
-            for line in netlib.utils.hexdump(data):
+            for line in strutils.hexdump(data):
                 self("\t%s %s %s" % line)
         else:
             for i in strutils.clean_bin(data).split("\n"):
