+++ test/netlib/test_strutils.py
@@ -38,3 +38,26 @@ def test_escaped_str_to_bytes():
     assert strutils.escaped_str_to_bytes(u"\\x08") == b"\b"
     assert strutils.escaped_str_to_bytes(u"&!?=\\\\)") == br"&!?=\)"
     assert strutils.escaped_str_to_bytes(u"ü") == b'\xc3\xbc'
+
+
+def test_isBin():
+    assert not strutils.isBin("testing\n\r")
+    assert strutils.isBin("testing\x01")
+    assert strutils.isBin("testing\x0e")
+    assert strutils.isBin("testing\x7f")
+
+
+def test_isXml():
+    assert not strutils.isXML("foo")
+    assert strutils.isXML("<foo")
+    assert strutils.isXML("  \n<foo")
+
+
+def test_clean_hanging_newline():
+    s = "foo\n"
+    assert strutils.clean_hanging_newline(s) == "foo"
+    assert strutils.clean_hanging_newline("foo") == "foo"
+
+
+def test_hexdump():
+    assert list(strutils.hexdump(b"one\0" * 10))
